lib/syscall_shim/arch/arm64: Add `uk_syscall_ctx` saving prologue
Implement an ARM inline assembly macro `UK_SYSCALL_USC_PROLOGUE_DEFINE`
function named after `pname` that switches to the auxiliary stack,
and starts pushing and storing the caller's context before passing it
as argument to a function whose name is defined by the second argument,
`fname`.

Althrough we supposedly don't touch the callee preserved registers in
this function, fully optimized images may inline the `fname` function
and end up messing the ABI. Therefore, make double sure this is avoided
by restoring the callee preserved registers before switching back to the
initial stack and returning.

To further emphasize to the compiler that we do not want anything
else in this function other than the prologue, optimize it with O3.

Notice that we make use of `TPIDRRO_EL0` as a scratch register so that
we can temporarily preserve x0's value. This should be safe as the
application/user themselves cannot change this register's value,
which means this register's value will always be known by us and we
could restore it anytime.

Some example OSes making use of TPIDRRO_EL0 would include Zephyr using
use this register for holding its `struct __cpu`, Windows using it to
hold the current CPU number, or Linux which uses it as both a scratch
register and a secondary thread ID.

Signed-off-by: Sergiu Moga <sergiu@unikraft.io>
Reviewed-by: Michalis Pappas <michalis@unikraft.io>
Approved-by: Razvan Deaconescu <razvand@unikraft.io>
GitHub-Closes: #1175