lib/syscall_shim/arch/x86_64: Add `struct uk_syscall_ctx` restore
Introduce an assembly written method `uk_syscall_ctx_popall` that
may be used in conjunction with a context that may be saved through a
`struct uk_syscall_ctx` structure. This method assumes that `%rsp`
points to the aforementioned structure and, furthermore, makes the
assumption that it is already aligned in memory to what this structure
should be (i.e. alignment of ECTX). Furthermore, and most importantly,
this method assumes that `gs_base` is set to the current LCPU's
correspondent `struct lcpu` pointer as it ends in a `swapgs`.

`uk_syscall_ctx_popall` simply disables IRQ's and restores the
extended context, userland context (`gs_base`/`fs_base` at the moment)
and the general purpose registers only to then finish with a `swapgs`
to swap Unikraft/Application `gs_base`s and an `iretq` to "teleport"
in the expected context.

Signed-off-by: Sergiu Moga <sergiu@unikraft.io>