lib/syscall_shim/arch/arm64: Add `struct uk_syscall_ctx` restore
Introduce an assembly written method `uk_syscall_ctx_popall` that
may be used in conjunction with a context that may be saved through a
`struct uk_syscall_ctx` structure. This method assumes that `sp`
points to the aforementioned structure and, furthermore, makes the
assumption that it is already aligned in memory to what this structure
should be (i.e. alignment of ECTX).

`uk_syscall_ctx_popall` simply disables IRQ's and restores the
extended context, userland context (`tpidr_el0` only at the moment)
and the general purpose registers only to `eret` to "teleport"
to the expected context.

Signed-off-by: Sergiu Moga <sergiu@unikraft.io>