lib/syscall_shim: Save `struct uk_syscall_ctx` in `uk_syscall6_r`
Being the center of all system call dispatching, it may be useful
to let `uk_syscall6_r` receive a `struct uk_syscall_ctx` as its
argument, containing the full context of the caller/parent.

Thus, if a system call has its `_u_` variant defined, that will be
the one that is used instead and the argument passed to it is
the aforementioned `struct uk_syscall_ctx` structure.

Signed-off-by: Sergiu Moga <sergiu@unikraft.io>
Reviewed-by: Michalis Pappas <michalis@unikraft.io>
Approved-by: Razvan Deaconescu <razvand@unikraft.io>
GitHub-Closes: #1175