lib/vfscore: Workaround for initrd extract
This commit introduces an option that activates a workaround in the event
that an `initrd` with "extract" was specified as the root file system via
the fstab feature. Because mounting would fail because there is no
filesystem mounted to where the archive can be extracted to, a "ramfs"
volume is mounted before extraction.

Checkpatch-Ignore: LEADING_SPACE
Signed-off-by: Simon Kuenzer <simon@unikraft.io>