lib/ramfs: Return proper errors on `ioctl` calls
Currently ramfs returns EINVAL for any ioctl request, which leads to the
Ruby binary compatibility application to fail, since it tries to set
`O_ASYNC` and breaks on error, even if it does not directly depends on it
being set.

Add a special case to treat `O_ASYNC`, return `ENOTSUP` by default, and
return `ENOTTY` on every `ioctl()` call related to a terminal.
The implementation is similar to what we currently have in `lib/9pfs/`,
introduced by #1098

GitHub-Fixes: #1186
Signed-off-by: Stefan Jumarea <stefanjumarea02@gmail.com>