lib/vfscore: Use `unsigned long` for `fcntl`'s third argument's type
Although `fcntl` is most frequently used in conjunction with `int`s
as its third argument, the use case involving file locking with
operations such as `F_SETLK` or `F_GETLK` may require that the
third argument is a pointer.

Therefore, allow for correct addresses above 2G to be passed through
`fcntl` by changing the type of the third argument to `unsigned long`.

Signed-off-by: Sergiu Moga <sergiu@unikraft.io>