lib/posix-unixsocket: Fix internal pipe leak
Previously stream sockets of a socket pair had the write end of their
internal pipe duplicated as a "backdoor pipe", used for non-connection
oriented I/O. This functionality is only desired for datagram sockets
and lead to an inadvertent leak of the write end, as well as missing
hangup events.
This change fixes this bug by restricting backdoor pipes to datagram
sockets only.

Signed-off-by: Andrei Tatar <andrei@unikraft.io>