lib/posix-unixsocket: Do not raise events on close
Previously a socket would raise the same set of hangup events on either
shutdown or when closing all open descriptors referring to it,
essentially because close would call shutdown internally. This lead to
the same hangup events being emitted in either case.
Linux, however, behaves differently, not raising any events on socket
close, an undocumented behavior which some brittle applications
inadvertently rely on.

This change bring posix-unixsocket's behavior in line with Linux,
raising hangup events only on shutdown and not on file cleanup.

Signed-off-by: Andrei Tatar <andrei@unikraft.io>