lib/posix-pipe: Fix stack overflow on struct init
The initialization of `struct pipe_node` in `uk_pipefile_create` could
previously cause a stack overflow due to the struct initializer being
stored on the stack, containing the entire pipe buffer.
This change prevents this failure case by ensuring that the pipe node is
initialized with assignments and memset.

Signed-off-by: Andrei Tatar <andrei@unikraft.io>
Reviewed-by: Michalis Pappas <michalis@unikraft.io>
Approved-by: Razvan Deaconescu <razvand@unikraft.io>
GitHub-Closes: #1228