plat/common/lcpu: Fix re-def warnings of `LCPU_AUXSP_OFFSET`
Add a temporary fix for the re-definition of the `LCPU_AUXSP_OFFSET`
macro warnings. This is caused by `syscall_prologue.h` architecture
specific files already having their own local temporary definition
of the macro due to it not being a good idea to include the very
header that defines this macro because it would create unwanted
dependencies throughout libraries that include `uk/syscall.h`.
Thus, in the case where `uk/syscall.h` is included before
`uk/plat/common/lcpu.h`, a warning related to the re-definition
of the macro is issued.

This must be fixed ASAP as part of the plat re-architecting track
where the LCPU subsystem is moved to its own driver/library and thus
getting rid of the weird dependencies. Until then, unfortunately,
care must be taken when keeping track of `LCPU_*` macro changes as
these might also have to be done in `syscall_prologue.h`.

Signed-off-by: Sergiu Moga <sergiu@unikraft.io>