lib/posix-poll: Fix select on files w/o epoll support
Previously calling select with an unpollable file one of the fd_sets
would immediately propagate the error from epoll. This is contrary to
behavior on Linux, where unpollable files in the read or write sets are
immediately returned as available, and ones in the except set are ignored.
This change corrects this behavior and brings it in line with Linux.

Checkpatch-Ignore: SPACING
Signed-off-by: Andrei Tatar <andrei@unikraft.io>