lib/vfscore: Stub `F_GETLK` to always return `F_UNLCK`
Since we always run in an exclusive virtual filesystem mode (no multiple
processes trying to acquire the lock on the same file), it is fine for
now to stub `F_GETLK` to always return an unlocked file through the
`F_UNLCK` flag.

This way, applications that do assume non-exclusive virtual filesystem
mechanisms can still proceed as if the file is free to be acquired.

Signed-off-by: Sergiu Moga <sergiu@unikraft.io>