lib/vfscore: Refactor virtual "initrd" fs driver towards "extract"
The commit refactors the virtual filesystem "initrd" to "extract".
This is done in order to make the archive extraction step of an initrd
more visible and separate from filesystem mounting. "extract" is
not mounting an underlying ramfs volume. This must be explicitly done
by a fstab entry which provides greater flexibility in setting up
the vfs layout.
At the moment "extract" supports only cpio archives and uses the
source device field to distinguish between an embedded archive and
an initrd module provided by a bootloader.

Checkpatch-Ignore: AVOID_EXTERNS
Signed-off-by: Simon Kuenzer <simon@unikraft.io>
Reviewed-by: Alexander Jung <alex@unikraft.io>
Reviewed-by: Razvan Deaconescu <razvand@unikraft.io>
Reviewed-by: Stefan Jumarea <stefanjumarea02@gmail.com>
Approved-by: Razvan Deaconescu <razvand@unikraft.io>
GitHub-Closes: #1262