lib/vfscore: Automount scheme with 3 tables
This commit introduces a fairly flexible mounting mechanism by
using three filesystem tables:
1) compiled-in, configured via menuconfig
2a) user-provided, read entries from kernel parameters ("vfs.fstab")
2b) fall-back, similar to "compiled-in"

If auto-mounting is activated, entries of the comiled-in table (1) are
mounted first. Afterwards, compiled-in fall-back entries (2b) are only
processed if the user did not provide any mount entries via the kernel
command line (2a). The compiled-in tables can be empty.
Note that each of the compiled-in tables are limited to four entries. The
size of the user-provided table can be configured.

Checkpatch-Ignore: LONG_LINE_STRING
Checkpatch-Ignore: LINE_SPACING
Signed-off-by: Simon Kuenzer <simon@unikraft.io>