{plat,lib,arch}/: Align auxstack allocation with regular stack
When first introduced, the auxiliary stack would be allocated through
`ukplat_auxsp_alloc`, a method that would directly return the value of
the final stack pointer, not the start of the stack buffer (stack grows
downwards in our usecases). Meanwhile, regular stacks would be allocated,
especially during thread creation, by first allocating the buffer and
only then assigning the thread's stack pointer to the result of the
`ukarch_gen_sp` operation on the aforementioned buffer. This
misalignment would see itself manifest through the fact that
`struct uk_thread`'s `struct _mem`'s `auxstack` field would end
up having the wrong value of the stack pointer instead of the
actual start of the buffer (the other end of the stack).

Although in practice this execution branch would never end up being
used to cause issues, whenever a thread would be deallocated, if
the previously mentioned `_mem.auxstack` and `_mem.auxstack_a` fields
are not initialized, the auxiliary stack is usually not deallocated as
it is expected of the user to do so, similarly to how it is done with
calls to `clone()`. To fix this, if the `auxsp` argument is 0, make sure
to initialize its related `struct uk_thread` fields are also initialized
to default values.

Lastly, as another change in the direction of alignment with regular
stacks, rename and move the auxiliary stack size configuration option
accordingly:
- `CONFIG_UKPLAT_AUXSP_PAGE_ORDER` -> `CONFIG_AUXSTACK_SIZE_PAGE_ORDER`
- move configuration option to `arch/Config.uk`

Signed-off-by: Sergiu Moga <sergiu@unikraft.io>