lib/uksched, lib/ukboot: Use `uk_memalign` for auxstack
For now freed threads wrongly try to free the other end of the stack
instead of the beginning of the allocated buffer.
If `CONFIG_LIBUKVMEM` is enabled this is even
worse since a `uk_free` is not enough and the VMA needs to be unmapped
for which we need to keep track of the length. Therefore, further
adjustments to `lib/uksched` need to be done before we can freely
allocate auxiliary stacks this way.

With this commit, `ukplat_auxsp_alloc` remains unused.

Please note that this commit is a temporary quick fix for a bug that
manifests whenever a thread gets free'd.

Signed-off-by: Sergiu Moga <sergiu@unikraft.io>
Reviewed-by: Sergiu Moga <sergiu@unikraft.io>
Reviewed-by: Razvan Deaconescu <razvand@unikraft.io>
Reviewed-by: Simon Kuenzer <simon@unikraft.io>
Approved-by: Razvan Deaconescu <razvand@unikraft.io>
GitHub-Closes: #1266