lib/uksched: Introduce UK_THREADF_AUXSP
Introduces a thread attribute flag that indicates that a thread has a
auxiliary stack available. For example, this stack is currently used
for handling binary system calls on a separate stack. In the future
there will be more use of this stack so that threads should be
classified depending on code execution that relies on the auxiliary
stack to be available.

Checkpatch-Ignore: SPLIT_STRING
Signed-off-by: Simon Kuenzer <simon@unikraft.io>