lib/uksched, lib/ukboot: Derive auxiliary stack for init thread
Instead of initializing an auxiliary stack for the "init" thread, we
derive the auxiliary stack from the current CPU context during scheduler
initialization. This simplifies the boot code because for the "init"
context we do not need to distinguish if we compile with scheduling or
without, we have only one case to consider: We need to initialize the
auxiliary stack before and independently of scheduling. This will allow
us to move scheduling initialization into the boot inittab at some point.

Signed-off-by: Simon Kuenzer <simon@unikraft.io>
Reviewed-by: Sergiu Moga <sergiu@unikraft.io>
Reviewed-by: Razvan Deaconescu <razvand@unikraft.io>
Reviewed-by: Simon Kuenzer <simon@unikraft.io>
Approved-by: Razvan Deaconescu <razvand@unikraft.io>
GitHub-Closes: #1266