lib/vfscore: Fix getcwd to use the kernel API
Previously the getcwd syscall behaved according to the libc API instead
of the kernel one, returning a pointer to the buffer containing the
path, as well as potentially allocating memory.
This change fixes this, providing the correct API: userspace is solely
responsible for memory allocation, and the syscall returns the length of
the output path.

Signed-off-by: Andrei Tatar <andrei@unikraft.io>
Reviewed-by: Stefan Jumarea <stefanjumarea02@gmail.com>
Reviewed-by: Simon Kuenzer <simon@unikraft.io>
Approved-by: Simon Kuenzer <simon@unikraft.io>
GitHub-Closes: #1268