lib/ukargparse: Introduce `uk_nextarg()`, `uk_nextarg_r()`
The functions `uk_nextarg()` and `uk_netxarg_r()` are helper functions
to process C-strings containing a list of arguments where the arguments
are separated by a special character (e.g., colon-separated).
The function is intended to be used to iterate over such arguments
strings.

This commit also adopts `lib/vfscore` to use the functions for fstab
parsing and `lib/uknetdev` to use the functions for kernel parameter
parsing. This is done for code deduplication.

Signed-off-by: Simon Kuenzer <simon@unikraft.io>