support/scripts: Auto-split words and disable info message
Trying to use readarray or anything else to split everything in one go
is not possible as the input file has both '\n' and ' ' delimiters.

Remove quotes and ignore the message as it is an info check.
This check would not have failed in the action even without the ignore.

Signed-off-by: Cezar Craciunoiu <cezar.craciunoiu@gmail.com>