lib/ukargparse: Introduce `uk_strnkeycmp()`
`uk_strnkeycmp()` is a helper function for parsing key-value pairs
that are represented in a C-string. The function returns with success
if the keyword matches with the one in a given C-string. In such a
case, it returns the byte offset of the beginning of the value. This
information can be used for further processing.

Signed-off-by: Simon Kuenzer <simon@unikraft.io>
Reviewed-by: Sergiu Moga <sergiu@unikraft.io>
Approved-by: Razvan Deaconescu <razvand@unikraft.io>
GitHub-Closes: #1282