lib/vfscore: Add `ifinitrd0`, `ifnoinitrd0` to `ukopts` (fstab)
This commit introduces the Unikraft fstab options `ifinitrd0` and
`ifnotinitrd0`. These options make an fstab entry conditional to the case
if an initrd module exists and was loaded by the bootloader or virtual
machine monitor (VMM).
This can be used to hide or only activate certain fstab entries on the
presence or absence of an initrd.

Signed-off-by: Simon Kuenzer <simon@unikraft.io>