lib/syscall_shim: Introduce `uk_sysrettab`
Implement a basic in-image table of routines to be executed when
exiting Unikraft context system calls. Unikraft System Call return
table is a section, `.uk_sysrettab`, similar to `.uk_inittab`, that
allows one to register priority-based routines into with the help of
the `uk_sysretcall_prio` macro-definition.

On exit, every system call tries to iterate over this list
of registered routines only if it is the first system call in the
call chain, i.e. if an application calls a system call that somehow,
through its implementation, ends up calling other system calls,
it is guaranteed that the routines are executed only when exiting the
first called system call. This is done with the help of a newly added
per-thread variable, `uk_syscallchain_count`, which acts similarly to
a refcount, by being incremented on every system call being
called and decremented whenever exiting such system call. Before being
decremented, the aforementioned table shall be iterated upon only
if this very variable's value is equal to 1, i.e. if it is iterated
upon on the exit from the first system call called in the call chain.

Notice how it is enforced that this per-thread variable is checked for
to be equal to `1` in the context of a binary system call. This is
because it should always be only one binary system call called in the
current call-chain - if there are more, then it is on us that somehow
Unikraft ended up invoking the binary system call handler through
instructions like `syscall`/`svc #0`, which is non-sense and
should be caught instantly.

Signed-off-by: Sergiu Moga <sergiu@unikraft.io>