lib/syscall_shim: Register `strace` through `.uk_sysrettab`
Since the `strace`-like printing is done at the end of every binary
system call invoked, we may as well register it into the system
call return table. In order to keep it being run only on exit of
binary system calls, add a check for the `struct uk_syscall_ctx`
field of `struct uk_sysrettab_ctx` - if we are exiting a native
system call, then it will always be `NULL`, otherwise, it will
have the value of the application's context previous to entering
the Unikraft context.

Signed-off-by: Sergiu Moga <sergiu@unikraft.io>