lib/ukboot: Add thread self-kill on shutdown through `.uk_sysrettab`
Whenever the shutdown signal would be issued, current application
threads may still not be made aware of this. This could be mean
trouble if we want a clean shutdown. Consider the following example
following a recevied shutdown IRQ:
- we signal the system that a shutdown request has been issued by
setting a global variable and then we return from the IRQ
- we wait for the application to do some system call that puts the
application thread to sleep
- the shutdown thread gets scheduled
- the fdtab destructor kicks in and closes some file descriptors, which
can make the shutdown thread go to sleep and result in a reschedule of
an application thread, unaware of shutdown
- the application thread resumes execution which might make it use file
descriptors that are no longer valid, resulting in an erroneous system
state

We fix this with the help of a shutdown handling hook registered in
`.uk_sysrettab`, that simply checks whether a shutdown signal has
been issued or not - if yes, then self-terminate current thread and
reschedule.

NOTE: A crucial unintuitive step for making this work is the need
to make sure that if one of the termination functions' implementation
somehow ends up calling a native system call, it will not result in
the premature self-termination of the shutdown thread once that
system call is exited.
We do this by setting the value of shutdown thread's
`uk_syscallchain_count` to `1` early on. This way, we are guaranteed
to never execute the `.uk_sysrettab` routines on the shutdown thread,
simply because the aforementioned variable will always be more than
`1` on `.uk_sysrettab` iteration attempts.

Signed-off-by: Sergiu Moga <sergiu@unikraft.io>