lib/posix-socket: Extend socket events
This commit extends the socket events so that each socket
keeps track of its current state. This allows event handlers
to determine when a socket function is called multiple times in
a row (e.g., nginx calls listen twice) and, for what type of
socket a close is performed (e.g., listen socket).

As part of this extension, the single socket event is split
up into dedicated event objects per event. This avoids calling
the hook for events that are not of interest.

Signed-off-by: Marc Rittinghaus <marc.rittinghaus@unikraft.io>
Reviewed-by: Simon Kuenzer <simon@unikraft.io>
Approved-by: Simon Kuenzer <simon@unikraft.io>
GitHub-Closes: #1279