lib/posix-socket: Introduce CREATE, BIND events
The current set of events does not allow UDP connections to be tracked as
they do not require a 'listen' call. UNIX domain sockets cannot be tracked
either, as they can work even without a `bind` operation. Because of all
this, this commit introduces two more state change events that can be
observed:
 - CREATE: A socket got created
 - BIND: An address got bind to a socket

Signed-off-by: Simon Kuenzer <simon@unikraft.io>
Approved-by: Razvan Deaconescu <razvand@unikraft.io>
Reviewed-by: Stefan Jumarea <stefanjumarea02@gmail.com>
Reviewed-by: Maria Sfiraiala <maria.sfiraiala@gmail.com>
GitHub-Closes: #1296