lib/vfscore: Prevent superfluous slashes in working directory
The mount point can contain a trailing slash (for example the root mount
point has a path of "/"). The path that is concatenated to the mount point
also contains a slash at the beginning. Therefore, extra care must be taken
to prevent two slashes from appearing besides each other.

Signed-off-by: Marco Schlumpp <marco@unikraft.io>
Reviewed-by: Stefan Jumarea <stefanjumarea02@gmail.com>
Reviewed-by: Radu Nichita <radunichita99@gmail.com>
Approved-by: Razvan Deaconescu <razvand@unikraft.io>
GitHub-Closes: #1302