lib/posix-poll: Fix EPOLL_CTL_MOD missing events
Previously a epoll_ctl(EPOLL_CTL_MOD) call that added new events to be
monitored would miss events already present on the target file.
This was due to a design error misunderstanding the semantics.

This change fixes this issue by making MOD behave as a pair of DEL & ADD
operations performed atomically on the epoll file, while minimizing
redundant work.

Signed-off-by: Andrei Tatar <andrei@unikraft.io>