arch/x86: Use function pointers instead of switch cases for ECTX ops
During the initialization of the extended context we would, depending
on the result of CPUID, set alignment, size and corresponding
store/restore methods of this context. The latter, we would set to the
value of an enum which we would afterwards use to decide which of the
methods to use when storing/restoring. Change this so that we
directly invoke such methods instead doing a switch case.

Signed-off-by: Sergiu Moga <sergiu@unikraft.io>