{arch,plat}/x86: Refactor FS_BASE/GS_BASE operations
Extend the GS_BASE read/write functionality to the FS_BASE register
and move all of it to libcontext. Additionally, make the initialization
of these routines, which depends on FSGSBASE being available, registered
as a Unikraft constructor. Because of this, generalize the initialization
of (KERNEL_)GS_BASE during LCPU architectural initialization, which
happens earlier, to use MSRs instead.

Signed-off-by: Sergiu Moga <sergiu@unikraft.io>
Co-authored-by: Andrei Tatar <andrei@unikraft.io>