lib/vfscore: Ensure we return symlink target on `namei_resolve`
The `namei_resolve` method is different from its `*_no_follow` variants
as in it is supposed to follow the symbolic links and eventually return
the dentry of the final symlink target. However, if the first dentry
hit in the cache is a symlink, we no longer check if it is a symlink
or not and simply return it. To fix this, simply add a check: if current
dentry pointer is a symlink, jump to the symlink logic.

Signed-off-by: Sergiu Moga <sergiu@unikraft.io>
Reviewed-by: Michalis Pappas <michalis@unikraft.io>
Approved-by: Razvan Deaconescu <razvand@unikraft.io>
GitHub-Closes: #1320