lib: Introduce generic (page guarded) stack allocator library
Implement a generic stack allocation wrapper over a given Unikraft
memory allocator. The library receives as arguments an allocator and,
if libukvmem is enabled, a virtual address space and an initial size.
The library works in three ways:
1. If libukvmem is disabled, the library will make use of the allocator
received as argument to allocate the initial allocator structure and
Unikraft compliant stacks.
2. If libukvmem is enabled, the library will make use of the allocator
received as argument to allocate the initial allocator structure and
Unikraft compliant stacks. The difference from 1. is that the virtual
address space and the initial size arguments are now available. The
two arguments will hint how much of the stack to premap whenever they
are allocated first, i.e. what is the initial stack size before
on-demand paging kicks in.
3. If libukvmem is enabled AND the configuration option of the library
integrating page guards is enabled, the allocator given as argument
will only be used once when allocating the initial allocator structure,
while the virtual address space and initial size arguments will be used
to create stack VMA's, specific to libukvmem.

With this addition, now libukboot is made aware of this stack allocator
and therefore a new configuration option is introduced to allow the
customization of the aforementioned initial size argument for when
libukvmem is enabled. Consequently, libukschedcoop's initialization
allows now for three additional allocator arguments for regular
stacks, for auxiliary stacks and for the TLS. This is done so that
libukboot can now initialize and pass to libukschedcoop separate
stack allocators, differentiated by the differently configured size,
alignment as well as the fact that, unlike regular stacks,
the auxiiliary stacks must always be fully mapped.

Signed-off-by: Sergiu Moga <sergiu@unikraft.io>