drivers/virtio/net: Ensure half-page alignment for netbuf
Currently there is a bug where if the net buffer (of size 2048) is split
into more than one descriptor the VMM reports the vring as full and
leading to us returning an error and wasting IRQs. The buffer would
be split in multiple virtio descriptors if the scatter-gather list
notices its span crosses more than one page (i.e. the buffer is not
entirely contained in one page) since it is not guaranteed that the two
virtually contiguous pages are also contiguous physically. To avoid
having this happen make sure that the buffer is always contained in one
page by having it aligned to half a page: if the buffer is half a page
in size and it is also half a page aligned it is mathematically
guaranteed to be entirely contained in one single page.

Signed-off-by: Sergiu Moga <sergiu@unikraft.io>