arch: Add I/O-able exception handling trampolines ctx creation
Implement the ability to jump from exception handling context to an
I/O-able context from which one can return to the context that the
exception handler was supposed to return to.

Essentially by creating a return-chain into a function that creates
room on a given stack to store execution environment
(struct ukarch_execenv) which would then return to a custom entry which
would then return to a routine that would restore the aforementioned
execution environment, one is able to freely do I/O, use any registers.

The caller of ukarch_ctx_init_ehtrampoX API's will be given a context
that is guaranteed to run with IRQ's enabled and from which, upon
return, execution from the point the exception handler was supposed
to return to is resumed.

        ┌───────────────────┐
▲       │                   │
│       │      struct       │
│	│      execenv      │
│       │                   │
│       │                   │
│       ├───────────────────┤
│       │                   │
│       │  execenv_load(ee) │
│       │                   │
│       ├───────────────────┤
│       │ entry(ee, arg0,   │
│       │       arg1…) for  │
│       │    ehtrampo0,1,2… │
│       │    variants       │
│       │                   │
│       │                   │
│       ├───────────────────┤
│       │                   │
│       │ ehtrampo_.._store │
exec    │                   │
order   │                   │
        └───────────────────┘

Signed-off-by: Sergiu Moga <sergiu@unikraft.io>