{arch,plat}: Add current LCPU to trap context
Introduce `lcpu_get_current_in_except()`.
Assuming we are in an exception handler on an exception stack,
fetches the current LCPU pointer in such a manner that we do not
assume any register state. This is done by dividng current stack
pointer relative to the LCPU stack base to find out the current
LCPU id.

Add the current LCPU field to trap context and use aforementioned
method to fetch it.

NOTE: On ARM64 SVC traps we don't do this because we are on the
auxiliary stack not an exception stack.

Signed-off-by: Sergiu Moga <sergiu@unikraft.io>