lib/ukcpio: Overwrite destination if exists
This change adds logic to CPIO extraction that attempts to remove or
rename an existing destination path. Specifically:
- regular files and symlinks will be unlinked
- empty directories will be removed
- non-empty directories will be renamed to NAME.0

When extracting a directory on top of an existing directory, the latter
is not removed or renamed, and only has mode bits adjusted.

We choose renaming over recursive deletion because:
- replacing directories with other files should ideally be a rare event
- recursive deletion, while storage efficient, is nontrivial and costly
  to perform; renaming OTOH is fast but wasteful
- we value boot latency in Unikraft, thus picking rename
- this tradeoff should be revisited if/when either (1) we have efficient
  recursive directory removal or (2) we value storage footprint over
  boot latency

Checkpatch-Ignore: AVOID_EXTERNS
Checkpatch-Ignore: FUNCTION_ARGUMENTS
Checkpatch-Ignore: STRCPY
Checkpatch-Ignore: STRLCPY
Signed-off-by: Andrei Tatar <andrei@unikraft.io>