lib/ukvmem: Fix out of range `vma_end` found as last element of list
Consider the following case:
               +------------+
               |VMA to unmap|
               +------------+      next VMA
          +----------+           +----------+
          |mapped VMA|           |mapped VMA|
          +----------+           +----------+
The current state of the code would simply assign `vma_end` to the
next mapped VMA without checking if the end of the VMA to unmap is
within its range. Fix this by adding this check and ensure we do
not assign `vma_end` to the next VMA if it does not meet the
requirements.

Signed-off-by: Sergiu Moga <sergiu@unikraft.io>
Reviewed-by: Michalis Pappas <michalis@unikraft.io>
Approved-by: Simon Kuenzer <simon@unikraft.io>
GitHub-Closes: #1322