lib/ukboot: Replace NOSCHED with INITSCHED
This commit introduces the hidden configuration option
`CONFIG_LIBUKBOOT_INITSCHED` that indicates that scheduler initialization
is configured with `lib/ukboot`.
The previous check for `!CONFIG_LIBUKBOOT_NOSCHED` was not sufficient,
as the option is not available if `CONFIG_LIBUKBOOT_NOALLOC` is set. The
reason is the dependency expressed for the corresponding `choice` in
the libraries' `Config.uk`. This would incorrectly mean that a scheduler is
being initialized.

Signed-off-by: Simon Kuenzer <simon@unikraft.io>