lib/ukboot: Replace NOALLOC with INITALLOC
This commit introduces the hidden configuration option
`CONFIG_LIBUKBOOT_INITALLOC` that indicates that allocator initialization
is configured with `lib/ukboot`.
We do this in order to be inline with the scheduler configuration
(`Config.uk`) and to have the possibility to add dependencies to the
corresponding `choice` of the allocator initialization configuration in
the future.

Signed-off-by: Simon Kuenzer <simon@unikraft.io>